Fix np.int/np.float crash in sparse-precomputed path
These aliases were removed in NumPy 1.24; the sparse-precomputed code path
raised on any modern install. Replace with np.int32 / np.float64.

+++ featuremap/featuremap_.py
@@ -2502,8 +2502,8 @@ class FeatureMAP(BaseEstimator):
                 )
             if not np.all(X.diagonal() == 0):
                 raise ValueError("Non-zero distances from samples to themselves!")
-            self._knn_indices = np.zeros((X.shape[0], self.n_neighbors), dtype=np.int)
-            self._knn_dists = np.zeros(self._knn_indices.shape, dtype=np.float)
+            self._knn_indices = np.zeros((X.shape[0], self.n_neighbors), dtype=np.int32)
+            self._knn_dists = np.zeros(self._knn_indices.shape, dtype=np.float64)
             for row_id in range(X.shape[0]):
                 # Find KNNs row-by-row
                 row_data = X[row_id].data
